requirements.txt
jupyter
pandas
matplotlib
pylast
python-dotenv
tqdm
spotipy